f32 instead of f64

+++ src/color.rs
@@ -1,7 +1,7 @@
 use crate::{hit::Interval, v3::V3};
 
 /// Apply a linear to gamma transform for gamma 2
-fn linear_to_gamma(linear_component: f64) -> f64 {
+fn linear_to_gamma(linear_component: f32) -> f32 {
     if linear_component > 0.0 {
         linear_component.sqrt()
     } else {
@@ -15,7 +15,7 @@ impl Color {
     pub const WHITE: Color = Color::new(1.0, 1.0, 1.0);
     pub const BLACK: Color = Color::new(0.0, 0.0, 0.0);
 
-    pub const fn grey(v: f64) -> Color {
+    pub const fn grey(v: f32) -> Color {
         Color::new(v, v, v)
     }
